commonwealth, blueos_startup_update: move save_file and locate_file helpers to commonwealth
